requirements.txt
absl-py==2.3.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
bokeh==3.8.1
Bottleneck==1.6.0
Cartopy==0.25.0
cattrs==25.3.0
certifi==2025.11.12
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dask==2025.11.0
decorator==5.2.1
distributed==2025.11.0
donfig==0.8.1.post1
# Editable install with no version control (drjit==1.2.0)
-e ./mitsuba3/ext/drjit
executing==2.2.1
flatbuffers==25.9.23
flox==0.10.7
fonttools==4.61.0
fsspec==2025.12.0
gast==0.7.0
google-crc32c==1.7.1
google-pasta==0.2.0
grpcio==1.76.0
h5netcdf==1.7.3
h5py==3.15.1
hatch-fancy-pypi-readme==25.1.0
hatchling==1.28.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipydatawidgets==4.3.5
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
llvmlite==0.46.0
locket==1.0.0
lxml==6.0.2
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
# Editable install with no version control (mitsuba==3.7.1)
-e ./mitsuba3
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
# Editable install with no version control (nanobind==2.9.2)
-e ./mitsuba3/ext/nanobind
narwhals==2.13.0
nc-time-axis==1.4.1
netCDF4==1.7.3
numba==0.63.1
numbagg==0.9.3
numcodecs==0.16.5
numpy==2.3.5
numpy-groupies==0.11.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.17.0.29
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.9.86
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydap==3.5.8
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
pyproj==3.7.2
pyshp==3.0.3
python-dateutil==2.9.0.post0
pythreejs==2.4.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-cache==1.2.1
rich==14.2.0
scikit_build_core==0.11.6
scipy==1.16.3
seaborn==0.13.2
shapely==2.1.2
sionna-rt==1.2.1
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8
sparse==0.17.0
stack-data==0.6.3
tblib==3.2.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
toolz==1.1.0
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.3
trove-classifiers==2025.12.1.14
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.6.2
wcwidth==0.2.14
WebOb==1.8.9
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xarray==2025.12.0
xyzservices==2025.11.0
zarr==3.1.5
zict==3.0.0
zipp==3.23.0
